Added package private class GVRPhysicsWorldObject
This class was added to avoid repeating code common to all objects that need to be added to the world
